move editor components in separate component
